--- src/guard/login.guard.ts
@@ -1,32 +1,0 @@
-import { ExecutionContext, Injectable } from '@nestjs/common';
-import { Reflector } from '@nestjs/core';
-import { AuthGuard } from '@nestjs/passport';
-import { Observable } from 'rxjs';
-import ErrorMessage from 'src/common/enums/error.message';
-import UnauthorizedError from 'src/common/errors/unauthorizedError';
-import { IS_PUBLIC_KEY } from 'src/decorator/public.decorator';
-
-@Injectable()
-export class JwtGuard extends AuthGuard('jwt') {
-  constructor(private reflector: Reflector) {
-    super();
-  }
-
-  canActivate(context: ExecutionContext): boolean | Promise<boolean> | Observable<boolean> {
-    const isPublic = this.reflector.getAllAndOverride<boolean>(IS_PUBLIC_KEY, [
-      context.getHandler(),
-      context.getClass()
-    ]);
-    if (isPublic) {
-      return true;
-    }
-    return super.canActivate(context);
-  }
-
-  handleRequest<TUser = any>(err: any, user: any): TUser {
-    if (err || !user) {
-      throw err || new UnauthorizedError(ErrorMessage.TOKEN_UNAUTHORIZED_NOTFOUND);
-    }
-    return user;
-  }
-}
